assignments/week-4/marketing-insights-analyzer/requirements.txt
streamlit==1.40.0
openai==1.54.0
python-dotenv==1.0.1
pandas==2.2.0
plotly==5.24.0
matplotlib==3.9.0
seaborn==0.13.2
nltk==3.9.1
spacy==3.8.2
nrclex==4.0.0
altair==4.2.2